+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.0.4
 - Removed terminal use, instead using `child_process` package to handle shell commands
 - Now shows a nice vscode-native Error message when decryption/encryption cannot happen for some reason
+- Added `decrypt in-place` and `encrypt in-place` right-mouse-menu buttons
 
 # 0.0.3
 - Bugfixes